Update README.md
This is a Project is Originally  an Idea of Vincent Garcia and is basically a k-nearest neighbor algorithm (k-NN) which is a widely used machine learning algorithm used for both classification and regression. k-NN algorithms are used in many research and industrial domains such as 3-dimensional object rendering, content-based image retrieval, statistics (estimation of entropies and divergences), biology (gene classification), etc. The processing time of the kNN search still remains the bottleneck in many application domains, especially in high dimensional spaces. In this work, we try to address this processing time issue by performing the kNN search using the GPU. 


This project Infuses the Parallel Programming and jointly through the use of Machine Learning Algorithms for KNN finds out the nearest K neighbour...